downloader/requirements.txt
yt-dlp~=2023.12.30
youtube-title-parse~=1.0.0
python-socketio~=5.11.1
requests~=2.31.0
websocket-client~=1.7.0
Flask~=3.0.2
flask-cors~=4.0.0
gunicorn~=21.2.0
